Isolate `cuda::std::extent` forward declarations
