+++ libcudacxx/include/cuda/std/__fwd/mdspan.h
@@ -28,7 +28,7 @@
 #  pragma system_header
 #endif // no system header
 
-#include <cuda/std/__cstddef/types.h>
+#include <cuda/std/__fwd/extents.h>
 #include <cuda/std/__type_traits/void_t.h>
 
 #include <cuda/std/__cccl/prologue.h>
@@ -106,14 +106,6 @@ inline constexpr bool __is_cuda_std_layout_left_or_right_mapping_v =
 
 // template<>
 // inline constexpr bool __is_cuda_std_layout_left_mapping_v<layout_left_padded> = true;
-
-template <class _IndexType, size_t... _Extents>
-class extents;
-
-template <class _Tp>
-inline constexpr bool __is_cuda_std_extents_v = false;
-template <class _IndexType, size_t... _Extents>
-inline constexpr bool __is_cuda_std_extents_v<extents<_IndexType, _Extents...>> = true;
 
 _CCCL_END_NAMESPACE_CUDA_STD
